shared/auth-lib-python/requirements.txt
aiohttp>=3.9.0
PyJWT>=2.8.0
fastapi>=0.100.0
python-multipart>=0.0.6